requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.4
requests==2.31.0
tkinter-tooltip==2.0.0
